Made authenticator into an interface and created a factory for it
